+++ client/components/planning-editor-standalone/profile.ts
@@ -1,192 +1,8 @@
 import {OrderedMap} from 'immutable';
 
-import {
-    IAuthoringFieldV2,
-    ICommonFieldConfig,
-    IContentProfileV2,
-    IDateTimeFieldConfig,
-    IDropdownConfigVocabulary,
-    IEditor3Config,
-    IVocabularyItem,
-} from 'superdesk-api';
-import {superdeskApi} from '../../superdeskApi';
-import {
-    IAttachmentsFieldConfig,
-} from '../../planning-extension/src/authoring-react-fields/planning-attachments/interfaces';
-import {getCustomVocabularyFields} from './field-adapters/custom-vocabularies';
+import {IContentProfileV2} from 'superdesk-api';
 import {getPlanningProfileFields} from './profile-fields';
-
-function getTextFieldConfig(options: {id: string; label: string, required: boolean}): IAuthoringFieldV2 {
-    const editor3ConfigWithoutFormatting: IEditor3Config = {
-        editorFormat: [],
-        minLength: undefined,
-        maxLength: undefined,
-        cleanPastedHtml: false,
-        singleLine: true,
-        disallowedCharacters: [],
-        showStatistics: false,
-        width: 100,
-    };
-
-    const field: IAuthoringFieldV2 = {
-        id: options.id,
-        name: options.label,
-        fieldType: 'editor3',
-        fieldConfig: {
-            ...editor3ConfigWithoutFormatting,
-            required: options.required,
-        },
-    };
-
-    return field;
-}
-
-function getDateTimeField(options: {id: string; label: string, required: boolean}): IAuthoringFieldV2 {
-    const config: IDateTimeFieldConfig = {
-        allowSeconds: false,
-    };
-
-    const field: IAuthoringFieldV2 = {
-        id: options.id,
-        name: options.label,
-        fieldType: 'datetime',
-        fieldConfig: {
-            ...config,
-            required: options.required,
-        },
-    };
-
-    return field;
-}
-
-export interface IFieldDefinition {
-    fieldId: string;
-    getField: (options: {required: boolean, id: string}) => IAuthoringFieldV2;
-    storageAdapter?: {
-        storeValue: <T extends IPlanningItem>(item: T, operationalValue: unknown) => T; // returns stored value
-        retrieveStoredValue:
-            <T extends IPlanningItem>(item: T, fieldId: string) => unknown; // returns operational value
-    };
-}
-
-type IFieldDefinitions = {[fieldId: string]: IFieldDefinition};
-
-export function getFieldDefinitions(): IFieldDefinitions {
-    const {gettext} = superdeskApi.localization;
-    const result: Array<IFieldDefinition> = [
-        {
-            fieldId: 'ednote',
-            getField: ({required, id}) => getTextFieldConfig({id: id, label: gettext('Ed Note'), required: required}),
-        },
-        {
-            fieldId: 'internal_note',
-            getField: ({required, id}) =>
-                getTextFieldConfig({id: id, label: gettext('Internal Note'), required: required}),
-        },
-        {
-            fieldId: 'name',
-            getField: ({required, id}) => getTextFieldConfig({id: id, label: gettext('Name'), required: required}),
-        },
-        {
-            fieldId: 'slugline',
-            getField: ({required, id}) => getTextFieldConfig({id: id, label: gettext('Slugline'), required: required}),
-        },
-        {
-            fieldId: 'description_text',
-            getField: ({required, id}) =>
-                getTextFieldConfig({id: id, label: gettext('Description'), required: required}),
-        },
-        {
-            fieldId: 'headline',
-            getField: ({required, id}) => getTextFieldConfig({id: id, label: gettext('Headline'), required: required}),
-        },
-        {
-            fieldId: 'planning_date',
-            getField: ({required, id}) =>
-                getDateTimeField({id: id, label: gettext('Planning date'), required: required}),
-        },
-        {
-            fieldId: 'files',
-            getField: ({required, id}) => {
-                const fieldConfig: IAttachmentsFieldConfig = {
-                    required,
-                };
-
-                const field: IAuthoringFieldV2 = {
-                    id: id,
-                    name: gettext('Attached files'),
-                    fieldType: 'files',
-                    fieldConfig: fieldConfig,
-                };
-
-                return field;
-            },
-        },
-        {
-            fieldId: 'place',
-            getField: ({id, required}) => {
-                const fieldConfig: IDropdownConfigVocabulary = {
-                    source: 'vocabulary',
-                    vocabularyId: 'locators',
-                    multiple: true,
-                    required: required,
-                };
-
-                const field: IAuthoringFieldV2 = {
-                    id: id,
-                    name: gettext('Place'),
-                    fieldType: 'dropdown',
-                    fieldConfig: fieldConfig,
-                };
-
-                return field;
-            },
-            storageAdapter: {
-                storeValue: (item, operationalValue: Array<string>) => {
-                    const vocabulary = superdeskApi.entities.vocabulary.getAll().get('locators');
-                    const vocabularyItems = new Map<IVocabularyItem['qcode'], IVocabularyItem>(
-                        vocabulary.items.map((item) => [item.qcode, item]),
-                    );
-
-                    return {
-                        ...item,
-                        place: operationalValue.map((qcode) => vocabularyItems.get(qcode)),
-                    };
-                },
-                retrieveStoredValue: (item, fieldId) => item[fieldId].map(({qcode}) => qcode),
-            },
-        },
-        {
-            fieldId: 'coverages',
-            getField: ({id, required}) => {
-                const fieldConfig: ICommonFieldConfig = {
-                    required,
-                };
-
-                const field: IAuthoringFieldV2 = {
-                    id: id,
-                    name: gettext('Coverages'),
-                    fieldType: 'coverages',
-                    fieldConfig: fieldConfig,
-                };
-
-                return field;
-            },
-        }
-    ];
-
-    result.push(
-        ...getCustomVocabularyFields(),
-    );
-
-    const resultObj = result.reduce((acc, item) => {
-        acc[item.fieldId] = item;
-
-        return acc;
-    }, {} as IFieldDefinitions);
-
-    return resultObj;
-}
+import {getFieldDefinitions} from './field-definitions/index';
 
 export function getProfile() {
     const planningFieldIds = getPlanningProfileFields();
